webdemo/requirements.txt
streamlit>=1.32.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
boto3>=1.39.9